requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.7.1
motor==3.4.0
python-dotenv==1.0.1
python-multipart==0.0.9